Add feature: users must register with a unique email address
When a user tries to submit a new registration form,
but use an email that is already registered,
they will be taken back the the form to try again.

Additionally, the form will be prepopulated with their previous attempt:
- name
- address
- city
- state
- zip code

But NOT their email or password.

Subsequently, all users who do not successfully complete the form will
now have this bonus experience - the form will be partly filled in
when there are other cases (ie. if a user leaves a field blank)