Pass user_params through a session instead of through a query string
Previously, the data was being passed through as a query string in the URL.
I think since the information is about a user's private details (address, etc)
It would be better to use a session so that information remains hidden from view.

Also, I deliberately delete the information from the session after it is used.