Add feature: Navigation Restrictions
Visitors, regular users, merchant employees and admins
will see a 404 error if they try to access a page that
requires an authorized user and they are unauthorized.

ApplicationController now has two helper methods:
- require_authorized_user
- logged_in_user

require_authorized_user is called with a before_action
in the merchant and admin dashboard controllers and in
the cart controller.

Each of the three controllers have a new private method
- unauthorized_user?
So that each implementation detail is encapsulated away
and each controller can customize who has permissions.