Add feature: Shopping Cart Buttons
When a visitor has items in their shoppping cart and visits the cart page
There are now buttons underneath each item to add/remove 1 from the qty.

The add button will disappear if the merchant does not have sufficient wares.
If an item is removed to a qty of zero, then the item itself will be removed from the cart.